Moved the track container classes from g4hough to trackbas_historic. Modified the relevant include paths.
