+++ simulation/g4simulation/g4hough/PHG4KalmanPatRec.C
@@ -8,16 +8,15 @@
 #include "PHG4KalmanPatRec.h"
 
 // g4hough includes
-#include "SvtxVertexMap.h"
-#include "SvtxVertexMap_v1.h"
-#include "SvtxVertex.h"
-#include "SvtxVertex_v1.h"
-#include "SvtxTrackMap.h"
-#include "SvtxTrackMap_v1.h"
-#include "SvtxTrack.h"
-#include "SvtxTrack_v1.h"
-#include "SvtxTrackState.h"
-
+#include <trackbase_historic/SvtxVertexMap.h>
+#include  <trackbase_historic/SvtxVertexMap_v1.h>
+#include  <trackbase_historic/SvtxVertex.h>
+#include  <trackbase_historic/SvtxVertex_v1.h>
+#include  <trackbase_historic/SvtxTrackMap.h>
+#include  <trackbase_historic/SvtxTrackMap_v1.h>
+#include  <trackbase_historic/SvtxTrack.h>
+#include  <trackbase_historic/SvtxTrack_v1.h>
+#include  <trackbase_historic/SvtxTrackState.h>
 #include <trackbase_historic/SvtxClusterMap.h>
 #include <trackbase_historic/SvtxCluster.h>
 #include <trackbase_historic/SvtxHitMap.h>
