+++ simulation/g4simulation/g4hough/PHG4SvtxMomentumRecal.C
@@ -1,6 +1,7 @@
 #include "PHG4SvtxMomentumRecal.h"
-#include "SvtxTrackMap.h"
-#include "SvtxTrack.h"
+
+#include  <trackbase_historic/SvtxTrackMap.h>
+#include  <trackbase_historic/SvtxTrack.h>
 
 // PHENIX includes
 #include <fun4all/Fun4AllReturnCodes.h>
