+++ simulation/g4simulation/g4hough/PHG4SvtxMomentumRecal.h
@@ -8,7 +8,7 @@
 //===========================================================
 
 #include "PHG4HoughTransform.h"
-#include "SvtxTrack.h"
+#include  <trackbase_historic/SvtxTrack.h>
 
 // PHENIX includes
 #include <fun4all/SubsysReco.h>
