+++ simulation/g4simulation/g4hough/PHG4SvtxTrackProjection.C
@@ -1,7 +1,8 @@
 #include "PHG4SvtxTrackProjection.h"
-#include "SvtxTrackMap.h"
-#include "SvtxTrack.h"
 #include "PHG4HoughTransform.h"
+
+#include  <trackbase_historic/SvtxTrackMap.h>
+#include  <trackbase_historic/SvtxTrack.h>
 
 // PHENIX includes
 #include <fun4all/Fun4AllReturnCodes.h>
